Bringing over vim settings from vim old repo.
This was done because I am going to try and use NVIM. These changes are
mostly identical to what's in my vim settings. There are some defaults that
Vim and neovim have that make a few of my old settings un needed. This is done
in a big chunk so if you want more information about each setting you can look
at the git history.

https://github.com/RyanHedges/dotvim