Add vim concerns to dotfiles
This might not actually be complete but this gets me really close to the setup I
had with vim. There are some concerns like plugins that this commit does not
address. This provides me with a simple starting point to get out of vim the
things that I am used to.