Add basic settings to zshrc
I did this for several reasons...

I want my be able to tab complete options efficiently with zsh and have the
first match in prompt so I can quickly hit enter.

I wanted to try and limit my history so that it only contains unique commands.
This should remove the need to hit up multiple times if I've ran the same
command multiple times in a row.

I also created my simple prompt. This isn't what I normally use, but it's a good
start as I read more about zsh and how it works, and how to configure it.