Install vim plugins using pathogen and git submodules
I did this because I want to add plugins to my vim set up. To accomplish that I
did several things. I installed vim pathogen to manage the plugins in a
centralized location and have them included at vim runtime.

For pathogen I decided to continue to use sub modules to manage the plugins.
This is why there is a .gitmodules directory added to the project.

I added a executable shell script for setting up vim. Currently it has a
dependency that you have brew installed for package managing and installs ack
and silver searcher. This is because my vimrc has a dependency on ack to use
silver searcher. This was a hard dependency before, but I changed it to use ack
and not silver searcher when silver search is not installed. This would make my
vimrc a little more usable on systems that have ack but not silver searcher.