De-couple from settop script
I did this because this script was using functions defined in the settop script
that I had used to drive out the script. I would like this script to be able to
be ran manually on it's own if needed so I decided to create the functions that
it needs.