Add styles for prompt for version control
I did this because I wanted to customize my prompt without the use of oh my zsh
or something like that. This is a built in library provided by zsh that knows
how to work with multiple types of version control. I am configuring this with
just git in mind but it should work for other VCS. I wrote comments trying to
describe what each option is doing but the documentation that I linked should
describe each option better if future me wants to edit this.