Include ruby version in right prompt
I did this because I want to know what the current ruby version is for
the project that I am in. This was a little harder to do because the
internet had a bunch of suggestions for creating functions. These
functions would not fetch the version as I changed directories. This has
something to do with `$()` vs `${}` and how they work together. I'll
need to look this up later.