Change cap A for tests to do documentation format
I did this because I no longer need the hanami test runner because the
transition to rails is complete. I have now changed this to run my
tests in documentation mode since most systems run tests using dots.
While debugging it is sometimes nice to see what order tests were
running or to quickly scan the descriptions around a failing tests to
potentially give a hint as to what is breaking.