small changes
This has some changes to git config that I don't remember fully doing.
At some point I did something with git lfs so yeah. I change the path in
zsh because I was having issues with installing an older postgresql for
a clients project. My path was originally `export
PATH="/usr/local/sbin:$PATH"` so that you can set it back when you are
reading this commit later. I also have vi for making changes to the
command line. It's interesting and I don't have a real problem with it
so I'm keeping it there. Also added a clear alias because why not!