Add NVM settings
Since many of my projects have requirements for NODE, add nvm to my
zshrc and update settop to install nvm if it's not available on the
system.

This includes helpers for nvm command line as well as function for
finding the current node version to use. Go to the nvm-sh project in
github for more directions on how to use.

ps-id: AFA9235D-4C2F-4C11-B12A-C82EAB3F47EC