Add auto detect and installation of ruby version using rbenv
I did this so that as I change into a directory it will check if there
is a .ruby-version file. If so it will compare that version to the
available versions. If the versions are not installed it will prompt me
to install the ruby version. I added colors so that I don't miss what's
going on since it requires me to agree to the prompt.