Remove nvim related items
I did this because the nvim related items I stopped configuring and have
not been using. I can add this back later when I need it. For now I'll
use Zed and it's implementation of Vim features as well as my current
set up for vim in a terminal for when I'm in a pinch.