Add gpg signing via ssh
This adds GPG signing of commits through the ssh format. You can read
about signing on the github documentation to make it all work together
with github:
https://docs.github.com/en/authentication/managing-commit-signature-verification/telling-git-about-your-signing-key#telling-git-about-your-ssh-key

I was concerned about the path for the signingkey, but I've decided that
I will use similar naming conventions across my machine where possible
and deal with times that those things cannot be the same.