Update zed/settings.json
This updates my zed configs in a couple ways.
First it creates a backup of the settings because there have been changes
to Zed that made them create a default backup. I'm committing it so that
if I need to I can refer to the backup. At some point in the future
I can remove this especially as the Zed editor evolves.

I've manually updated the Ruby language server list because there have
been some updates to it since I last made it. This has been formatted
through the automatic update to the settings.json

It automatically renames assistant to agent, probably for marketing
reasons because of AI hype.

I've added the new minimap feature to make sure it shows for my files.

I haven't had the best experience with Ruby where the LSP and other
things seem to fail fairly frequently and setting use bundler to true
seems like the "right" thing to do for getting Ruby to work well. I have
not seen the best results for either so it probably needs more dialing
in before I can call working with Ruby in zed finalized.

Vertical scroll margin "Control Editor scroll beyond the last line"
because I want to have some more space when scrolling while also not
letting it go to crazy. You can read about it here:
https://zed.dev/docs/visual-customization?highlight=vertical_scroll_margin#editor-scrollbar